telegram-bot/requirements.txt
boto3
Flask
python-telegram-bot
requests
paramiko